Update to bsnes v029 release.
A new version of bsnes has been released. It contains a few minor emulation fixes, as well as user interface improvements. Behind the scenes, the source has been cleaned up more in preparation for running the CPU and PPU (video processor) separately from each other (eg with no enslavement.) This is required for implementing a clock cycle based PPU renderer.
    - Greatly improved invalid DMA transfer behavior, should be nearly perfect now
    - Major code cleanup -- most importantly, almost all PPU timing-related settings moved back to PPU, from CPU
    - Added option to auto-detect file type by inspecting file headers rather than file extensions
    - Rewrote video filter system to move it out of the emulation core -- HQ2x and Scale2x will work even in hires and interlace modes now, 50% scanline filter added
    - Re-added bsnes window icon
    - Added new controller graphic when assigning joypad keys [FitzRoy]
    - Redundant "Advanced" panel settings which can be configured via the GUI are no longer displayed
    - Improved speed regulation settings
    - XP and Vista themes will now apply to bsnes controls
    - Added "Path Settings" window to allow easy selection of default file directories
    - Tab key now mostly works throughout most of the GUI (needs improvement)
    - Main window will no longer disappear when setting a video multipler which results in a window size larger than the current desktop resolution
    - Added two new advanced options: one to control GUI window opacity, and one to adjust the statusbar text

+++ src/lib/ruby/ruby.cpp
@@ -1,8 +1,7 @@
 #include <ruby/ruby.h>
+#include <ruby/ruby_impl.cpp>
 
 namespace ruby {
-
-#include <ruby/ruby.impl.h>
 
 VideoInterface video;
 AudioInterface audio;
@@ -73,7 +72,7 @@ void VideoInterface::term() {
 bool VideoInterface::cap(Video::Setting setting) { return p ? p->cap(setting) : false; }
 uintptr_t VideoInterface::get(Video::Setting setting) { return p ? p->get(setting) : false; }
 bool VideoInterface::set(Video::Setting setting, uintptr_t param) { return p ? p->set(setting, param) : false; }
-bool VideoInterface::lock(uint16_t *&data, unsigned &pitch) { return p ? p->lock(data, pitch) : false; }
+bool VideoInterface::lock(uint32_t *&data, unsigned &pitch) { return p ? p->lock(data, pitch) : false; }
 void VideoInterface::unlock() { if(p) p->unlock(); }
 void VideoInterface::clear() { if(p) p->clear(); }
 void VideoInterface::refresh(unsigned width, unsigned height) { if(p) p->refresh(width, height); }
